+++ integration-tests/forklift/setupTests.ts
@@ -1,4 +1,4 @@
-import { startForklift, stopForklift } from "./src/chopsticks"
+import { startForklift, stopForklift } from "./src/lib/forklift"
 
 export default async function setupTests() {
   await startForklift()
