+++ integration-tests/forklift/src/archive.spec.ts
@@ -1,7 +1,7 @@
 import { paseo } from "@polkadot-api/descriptors"
 import { createClient, HexString } from "polkadot-api"
 import { describe, expect, it } from "vitest"
-import { ALICE, getForkliftProvider } from "./chopsticks"
+import { ALICE, getForkliftProvider } from "./lib/forklift"
 
 describe("archive", () => {
   it("loads information from past blocks", async () => {
